refactor: replaced image urls with assets
