+++ src/components/Header.jsx
@@ -1,15 +1,18 @@
-import { signOut } from "firebase/auth";
 import { MdOutlineLogout, MdLogin } from "react-icons/md";
 import { useDispatch, useSelector } from "react-redux";
-import { Link, useNavigate } from "react-router-dom";
+import { Link, useLocation, useNavigate } from "react-router-dom";
 import { logout } from "features/user/userSlice";
 import { auth } from "../firebase/firebase";
+import { signOut } from "firebase/auth";
+import { defaultAvatar } from "utils/Constants";
+import logo from "assets/jigsaw.png";
 
 export default function Header() {
   const { user } = useSelector((state) => state.user);
   const currentUser = auth?.currentUser;
   const navigate = useNavigate();
   const dispatch = useDispatch();
+  const location = useLocation();
 
   const logoutHandler = async () => {
     dispatch(logout());
@@ -19,18 +22,19 @@ export default function Header() {
   };
 
   return (
-    <div className="p-3 bg-slate-100 font-semibold flex justify-between items-center">
-      <Link to="/">
-        <h1 className="text-gray-600 text-xl">SocialMedia</h1>
+    <div className="flex h-12 items-center justify-between bg-slate-100 p-3 font-semibold">
+      <Link to="/" className="flex h-full">
+        <img src={logo} alt="main-logo" className="aspect-auto" />
+        <h1 className="text-xl text-gray-600">Jigsaw</h1>
       </Link>
       <div className="flex gap-4">
         <Link to="/profile">
           <img
-            className="rounded-full max-h-9 w-fit object-cover aspect-square"
+            className="aspect-square max-h-9 w-fit rounded-full object-cover"
             src={
               currentUser?.photoURL
                 ? currentUser?.photoURL
-                : "http://cdn.onlinewebfonts.com/svg/img_264570.png"
+                : defaultAvatar.cover
             }
             alt="user-avatar"
           />
@@ -40,7 +44,7 @@ export default function Header() {
             <MdOutlineLogout className="text-2xl" />
           </button>
         ) : (
-          <Link to="/login" className="self-center">
+          <Link to="/login" className="self-center" state={{ from: location }}>
             <MdLogin className="text-2xl" />
           </Link>
         )}
